Use GitHub actions for CI
Before, the repository lived in the legacy `NewBamboo` organisation.

Since the repository has been migrated to the `thoughtbot` organisation,
it makes sense to move away from the old Travis configuration.

GitHub actions is a reasonable choice since it is easy to use, well
documented, and builds are visible directly in GitHub and automatically
associated to the thoughtbot GitHub organisation.