Refactor specs
This refactors the existing specs to bring them more inline with our typical
approach outlined in the [guides](thoughtbot/guides), leaning on the [Four Phase
Test](https://thoughtbot.com/blog/four-phase-test) and [Let's
Not](https://thoughtbot.com/blog/lets-not) approaches.

The objective of this refactor is to make the tests more readable and accessible
for new contributors.

A dependency on `rspec-its` is also removed.

The `belongs_to_static` helper method is also refactored to use `.find_by_id`
(which returns `nil` for missing records) instead of `.find` and rescuing
`RecordNotFound` errors.