Coerce #find_by_id argument to integer
Ensures a matching record is found if passing a string representation of
the id, avoiding the need to coerce the argument into a string
beforehand.

E.g. `MyRecord.find_by_id("1")` returns matching `record(id: 1)`.

If the string does not coerce into an integer then the original string
argument is used. To cater for use cases where non-numeric IDs are used
for records.